+++ src/components/movieList/SkeletonListItem.ts
